Update Razor to support Phoenix 1.3 upgrades in razor-phoenix
This is a corresponding change to
https://github.com/carbonfive/razor-phoenix/pull/27, which introduced the
Phoenix 1.3 context file structure. In order to support this file structure, we
need to add an extra rename command for an additional folder.

Additionally, we introduce a `-r` (repo) argument to allow the user to
specify a different repository for the prototype app. This is helpful in
debugging scenarios.